python/Module12/python-cognito-sam/cognito_function/requirements.txt
requests
python_jose